тимчасовий commit

+++ practic/src/main/java/com/group/practic/repository/ChapterRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChapterRepository extends JpaRepository<ChapterEntity, Integer> {
+public interface ChapterRepository extends JpaRepository<ChapterEntity, Long> {
 
 }
