+++ practic/src/main/java/com/group/practic/repository/LevelRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LevelRepository extends JpaRepository<LevelEntity, Integer> {
+public interface LevelRepository extends JpaRepository<LevelEntity, Long> {
 
     List<LevelEntity> findAllByCourse(CourseEntity course);
 }
