+++ practic/src/main/java/com/group/practic/repository/SubChapterRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubChapterRepository extends JpaRepository<SubChapterEntity, Integer> {
+public interface SubChapterRepository extends JpaRepository<SubChapterEntity, Long> {
 }
