+++ practic/src/main/java/com/group/practic/repository/SubSubChapterRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubSubChapterRepository extends JpaRepository<SubSubChapterEntity, Integer> {
+public interface SubSubChapterRepository extends JpaRepository<SubSubChapterEntity, Long> {
 }
